+++ utoipa-gen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog - utoipa-gen
 
+## Unreleased
+
+### Added
+
+* Add support for jiff v0.2 (https://github.com/juahku/utoipa/pull/1332)
+
 ## 5.3.1 - Jan 6 2025
 
 ### Fixed